Adding extension points to property directive resolver.
