+++ src/Framework/Framework/Compilation/ControlTree/IAbstractPropertyDeclarationDirective.cs
@@ -1,4 +1,5 @@
 ﻿using DotVVM.Framework.Compilation.Parser.Binding.Parser;
+using System.Collections.Generic;
 using System.Reflection;
 
 namespace DotVVM.Framework.Compilation.ControlTree
@@ -13,9 +14,11 @@ namespace DotVVM.Framework.Compilation.ControlTree
     {
         SimpleNameBindingParserNode NameSyntax { get; }
         TypeReferenceBindingParserNode PropertyTypeSyntax { get; }
+        BindingParserNode? InitializerSyntax { get; }
         ITypeDescriptor? PropertyType { get; }
         ITypeDescriptor? DeclaringType { get; set; }
         object? InitialValue { get; }
+        IList<IAbstractDirectiveAttributeReference> Attributes { get; }
     }
 
 }
